Change `index` to double

+++ lib/src/components/features_tour_minxin.dart
@@ -1,7 +1,7 @@
 import 'package:features_tour/src/models/instruction_result.dart';
 
 mixin FeaturesTourStateMixin {
-  int get index => throw UnimplementedError();
+  double get index => throw UnimplementedError();
 
   Future<IntroduceResult> showIntroduce(FeaturesTourStateMixin calledState) =>
       throw UnimplementedError();
